+++ packages/core/src/__tests__/mitosis.test.ts
@@ -1,4 +1,6 @@
 import { componentToMitosis } from '@/generators/mitosis';
+import { createMitosisComponent } from '@/helpers/create-mitosis-component';
+import { createMitosisNode } from '@/helpers/create-mitosis-node';
 import { runTestsForTarget } from './test-generator';
 
 describe('Mitosis, format: legacy', () => {
@@ -29,5 +31,36 @@ describe('Mitosis, format: react', () => {
     },
     target: 'mitosis',
     generator: componentToMitosis,
+  });
+});
+
+describe('Can encode <> in text', () => {
+  it('should encode <> in text', () => {
+    const result = componentToMitosis()({
+      component: createMitosisComponent({
+        children: [
+          createMitosisNode({
+            properties: { _text: '<>{}' },
+          }),
+        ],
+        hooks: {},
+      }),
+    });
+
+    expect(result).toMatchSnapshot();
+  });
+  it('should not encode valid jsx', () => {
+    const result = componentToMitosis()({
+      component: createMitosisComponent({
+        children: [
+          createMitosisNode({
+            properties: { _text: 'hello <b>world</b>' },
+          }),
+        ],
+        hooks: {},
+      }),
+    });
+
+    expect(result).toMatchSnapshot();
   });
 });