+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 0.7.2
+
+### Patch Changes
+
+- f87bd64: [Angular] Use tagName to generate
+
+## 0.7.1
+
+### Patch Changes
+
+- 3d44b65: Properly encode "<", ">", and other chars in text in Mitosis JSX output
+
 ## 0.7.0
 
 ### Minor Changes